View source word count issue fix and change apply transaction action method

+++ src/services/repository/DraftRepository.php
@@ -14,7 +14,9 @@ use Craft;
 use Exception;
 use craft\db\Query;
 use craft\elements\Entry;
+use yii\web\NotFoundHttpException;
 use acclaro\translations\Translations;
+use craft\errors\InvalidElementException;
 use acclaro\translations\models\FileModel;
 use acclaro\translations\records\FileRecord;
 
@@ -129,24 +131,36 @@ class DraftRepository
     public function applyTranslationDraft($fileId)
     {
         $file = Translations::$plugin->fileRepository->getFileById($fileId);
-        
+
+        $isGlobal = false;
         // Get file's draft
         $draft = Translations::$plugin->draftRepository->getDraftById($file->draftId, $file->targetSite);
+        if(empty($draft)) {
+            $isGlobal = true;
+            $draft = Translations::$plugin->globalSetDraftRepository->getDraftById($file->draftId, $file->targetSite);
+        }
+
         if (!$draft) {
             throw new NotFoundHttpException('Draft not found');
         }
 
         try {
-            // Let's try saving the element prior to applying draft
-            if (!Craft::$app->getElements()->saveElement($draft)) {
-                throw new InvalidElementException($draft);
+            if ($isGlobal) {
+                $newEntry = Translations::$plugin->globalSetDraftRepository->publishDraft($draft);
+            } else {
+                // Let's try saving the element prior to applying draft
+                if (!Craft::$app->getElements()->saveElement($draft)) {
+                    throw new InvalidElementException($draft);
+                }
+
+                // Let's remove the auto-propagated drafts
+                Translations::$plugin->draftRepository->deleteAutoPropagatedDrafts($file->draftId, $file->targetSite);
+
+                // Apply the draft to the entry
+                $newEntry = Craft::$app->getDrafts()->applyDraft($draft);
             }
-            
-            // Let's remove the auto-propagated drafts
-            Translations::$plugin->draftRepository->deleteAutoPropagatedDrafts($file->draftId, $file->targetSite);
-            
-            // Apply the draft to the entry
-            $newEntry = Craft::$app->getDrafts()->applyDraft($draft);
+
+
         } catch (InvalidElementException $e) {
             Craft::$app->getSession()->setError(Craft::t('app', 'Couldn’t publish draft.'));
             // Send the draft back to the template
